Run laptop script through CI
Make sure that the script itself does not get out of date by running
the the suite on monthly basis. Thus ensuring if packages are
deprecated or links become invalid that this script remains evergreen.

This does not attempt to replicate exactly how a user might run this
script. It ignores the concept of user and root user and just installs
things as root, though it does maintain the use of sudo.

The environment is also much more barebones than what would typically
be setup, probably for good, but some finagling needed to happen to
make this run properly.

This also setups a concept of base requirements in order to run this
script at all, these are defined as dependencies.

* Setup baseline for running this script
* Install required dependencies
* Fix script reference to programs, use packages instead
* Fix reference to support directory
* Configure keyboard
* Configure locales
* Allow github domain to not require ssh
* Build script monthly
* Do not require gnupg.gpg it's not used realistically
* Attempt to start postgres since it may not start from automatically
* Force update dotfiles