Update script to accomodate other machines
Switch to a generator/build system that provides a way for each
machine to specify. It's not very robust, but it works well enough.

This completely overhauls how this script operates. Rather than having
a few files handle the whole task this breaks out each task
individually.

It breaks this out by doing the following:

- Check that system has been correctly setup
- Install set of required packages
- Setup system to be able to install packages for machine
- Install packages for machine type
- Finish configurations for machine

This by separating out package lists into separate files and build
tasks in separate files.