Security key setup
The goal for this configuration is to be as hands off as
possible. Until now this meant setting up an ssh key, getting it to
the approriate servers and transferring my pgp key installed on the
machine was arduous. Adding this allowed me to plug in my security key
and currently only have to worry about adding sudo to the user
beforehand.

* Setup bashrc.local for easy sourcing and later removal if necessary
* Move around install and builds of pgp key until after more installs
have happened

Potential Issues:
* I don't have a really good way of testing this short of wiping a
machine. This could not work as expected.